Initial import
Import the initial code from my computer.

Initial code logs into the current Azure account and stores the context.
It then retrieves the tennant ID.
After the tenant ID is acquired it then takes the context and runs a filter search on the token cache and sorts it on the expiration time property and saves the output.