Fix bug
Fixed a bug with the message box system, only load an icon object if the messagebox is going to render. This eliminates an issue where the script would throw an error stating that the message box icon object was not found because teh assembly was not loaded. The assembly only loads if the message box is going to render. This also optimizes the code by taking out an unnecessary message box icon instanciation ifthe box is not needed.