Apply best practice
Switch the equality so that null is on the left side.
Had to make it so that equality was checking the first instance, not the array of null.
The reason it was failing before was an array of null is not equal to null as the array existed, therefore it failed teh script.
Now it checks the first row of data for the column and now the if statement can be best practice.