Add support for paging
Add support for paging of the Graph API query results. The page size is 100 in azure at the time of this writing, so it will continue to return users in stacks of 100.
Graph API based filtering will be added in a future version to make the queries run faster and more efficiently by reducing the number of returned items to a minimum instead of returning all users and filtering locally.

Add an auto filter for the excluded user GUID parameter so that it can automatically remove any whitespace passed in via the parameter.