Add Multi Principal and Permission Support
Add the ability to add multiple permissions to the specified principal.
But wait, there is more!
You can also apply the specified permissions to multiple principals at the same time! (multiple permissions are supported to be added to the specified principals if you so chose).
Added the ability to grant permissions that are not Graph API native (e.g. Defender for Endpoint).